backend/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coverage==5.5
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.1.8
django-cache-url==3.2.3
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-extensions==0.7.0
environs==8.0.0
httpie==2.4.0
idna==2.10
Markdown==3.3.4
marshmallow==3.11.1
psycopg2-binary==2.8.6
Pygments==2.8.1
PyJWT==2.0.1
PySocks==1.7.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.4
